Don't compile nodejs again if we'd be building the same version
Whenever we build nodejs, we create a file
(/usr/local/share/node_version) that stores the sha1 hash of the current
commit.  Whenever the recipe is run, we first run "git ls-remote" to see
if the remote hash matches (it might is the case of a head ref, but
needs to be dereferenced in the case of a tag ref).  Then we clone the
repo and check the hash against the checked out branch.  If either one
matches, we don't need to build.